# Move MVCCKey encoding into C++.