--- a/roachpb/internal.proto
+++ b/roachpb/internal.proto
@@ -119,6 +119,8 @@ message RaftSnapshotData {
   message KeyValue {
     optional bytes key = 1;
     optional bytes value = 2;
+    optional int64 wall_time = 3 [(gogoproto.nullable) = false];
+    optional int32 logical = 4 [(gogoproto.nullable) = false];
   }
   // The latest RangeDescriptor
   optional RangeDescriptor range_descriptor = 1 [(gogoproto.nullable) = false];