--- a/storage/engine/gc.go
+++ b/storage/engine/gc.go
@@ -58,12 +58,7 @@ func (gc *GarbageCollector) Filter(keys []MVCCKey, values [][]byte) roachpb.Time
 	delTS := roachpb.ZeroTimestamp
 	survivors := false
 	for i, key := range keys {
-		_, ts, isValue, err := MVCCDecodeKey(key)
-		if err != nil {
-			log.Errorf("unable to decode MVCC key: %q: %v", key, err)
-			return roachpb.ZeroTimestamp
-		}
-		if !isValue {
+		if !key.IsValue() {
 			log.Errorf("unexpected MVCC metadata encountered: %q", key)
 			return roachpb.ZeroTimestamp
 		}
@@ -77,8 +72,8 @@ func (gc *GarbageCollector) Filter(keys []MVCCKey, values [][]byte) roachpb.Time
 			}
 		}
 		// If we encounter a version older than our GC timestamp, mark for deletion.
-		if ts.Less(gc.expiration) {
-			delTS = ts
+		if key.Timestamp.Less(gc.expiration) {
+			delTS = key.Timestamp
 			break
 		} else if !deleted {
 			survivors = true
@@ -87,14 +82,7 @@ func (gc *GarbageCollector) Filter(keys []MVCCKey, values [][]byte) roachpb.Time
 	// If there are no non-deleted survivors, return timestamp of first key
 	// to delete all entries.
 	if !survivors {
-		_, ts, _, err := MVCCDecodeKey(keys[0])
-		if err != nil {
-			// TODO(tschottdorf): Perhaps we should be propagating an error
-			// (e.g. ReplicaCorruptionError) up to the caller.
-			log.Errorf("unable to decode MVCC key: %q: %v", keys[0], err)
-			return roachpb.ZeroTimestamp
-		}
-		return ts
+		return keys[0].Timestamp
 	}
 	return delTS
 }