--- a/storage/replica_command.go
+++ b/storage/replica_command.go
@@ -1105,9 +1105,10 @@ func (r *Replica) TruncateLog(batch engine.Engine, ms *engine.MVCCStats, h roach
 	}
 	start := keys.RaftLogKey(rangeID, 0)
 	end := keys.RaftLogKey(rangeID, args.Index)
-	if err = batch.Iterate(engine.MVCCEncodeKey(start), engine.MVCCEncodeKey(end), func(kv engine.MVCCKeyValue) (bool, error) {
-		return false, batch.Clear(kv.Key)
-	}); err != nil {
+	if err = batch.Iterate(engine.MakeMVCCMetadataKey(start), engine.MakeMVCCMetadataKey(end),
+		func(kv engine.MVCCKeyValue) (bool, error) {
+			return false, batch.Clear(kv.Key)
+		}); err != nil {
 		return reply, err
 	}
 	tState := roachpb.RaftTruncatedState{