Changed all of the Engine/Iterator interfaces to use MVCCKeys. Moving
the MVCCKey encoding from Go to C++ simplifies the usage of
MVCCKeys. The MVCCKey type is now a structure composed of a roachpb.Key
and a timestamp. This is both more convenient to use and a minor perf
win, though the motivation behind this change was to pave the way to
using a custom RocksDB comparator.

```
name                                old time/op    new time/op    delta
MVCCScan1Version1Row8Bytes-8          4.70µs ± 2%    4.46µs ± 1%   -5.01% (p=0.000 n=9+8)
MVCCScan1Version10Rows8Bytes-8        23.7µs ± 1%    22.6µs ± 0%   -4.95% (p=0.000 n=10+9)
MVCCScan1Version100Rows8Bytes-8        200µs ± 1%     190µs ± 1%   -4.81% (p=0.000 n=9+9)
MVCCScan1Version1000Rows8Bytes-8      1.94ms ± 0%    1.85ms ± 0%   -4.39% (p=0.000 n=9+10)
```

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/3337)
<!-- Reviewable:end -->